backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
qrcode==7.4.2
pillow==10.2.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1